% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readMsp.R
\name{.groupIndex}
\alias{.groupIndex}
\title{Construct group index}
\usage{
.groupIndex(breaks)
}
\arguments{
\item{breaks}{Vector of integers specifying the index location where a group terminates}
}
\value{
a vector of index values the same length as the vector used to generate breaks
}
\description{
Construct group index
}
