% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commentParser.R
\name{.parseCommentVector}
\alias{.parseCommentVector}
\title{Parse vector of comment strings}
\usage{
.parseCommentVector(comments, indexs, commentType)
}
\arguments{
\item{comments}{a character vector containing the comment fields}

\item{indexs}{the group index of each comment}

\item{commentType}{either `NIST` or `MoNA` - used for comment parsing}
}
\value{
a data.table
}
\description{
Parse vector of comment strings
}
