% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processHeader.R
\name{.parseHeaderDt}
\alias{.parseHeaderDt}
\title{Process header data.table containing comment string}
\usage{
.parseHeaderDt(headerDt, commentType = commentType)
}
\arguments{
\item{headerDt}{a data.table returned from .processHeader()}

\item{commentType}{either `NIST` or `MoNA` - used for comment parsing}
}
\value{
a data.table with additional parsed fields from the comment comment
  strings
}
\description{
Wrapper function that takes a data.table containing header
information and a `Comment` column full of text. The `Comment` column is
parsed and a data.table with additional columns is returned. The original
`Comment` column is removed.
}
