% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processHeader.R
\name{.processHeader}
\alias{.processHeader}
\title{Process the header from an msp file}
\usage{
.processHeader(header, groupIndex, commentType = commentType)
}
\arguments{
\item{header}{character vector containing the header lines only}

\item{groupIndex}{an integer vector the same length as `header` providing a
index value that links the `header` information to data chunks (spectral
data)}

\item{commentType}{either `NIST` or `MoNA` - used for comment parsing}
}
\value{
a fully formatted data.table
}
\description{
Top level header processor for msp files. Provide a character
  vector containing the header lines and corrisponding groupIndex.
}
