% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commentParser.R
\name{.splitByEquals}
\alias{.splitByEquals}
\title{Split name-value pairs by an equal sign}
\usage{
.splitByEquals(nameValuePairs)
}
\arguments{
\item{nameValuePairs}{a list of vectors containing name value pairs}
}
\value{
returns a named vector
}
\description{
Split name-value pairs by an equal sign
}
