% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{.splitIndex}
\alias{.splitIndex}
\title{Split a vector into groups}
\usage{
.splitIndex(index, nGroups, randomize = FALSE)
}
\arguments{
\item{index}{the vector of index values to be split into groups}

\item{nGroups}{number of groups to be generated}

\item{randomize}{should the groups contain a random or ordered sampling from
the index vector}
}
\value{
returns a list containing `groups` and `groupIndex` both of which are
  lists of length nGroups
}
\description{
Split a vector into groups
}
