% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastRead.R
\name{f_read}
\alias{f_read}
\title{Read a File}
\usage{
f_read(file)
}
\arguments{
\item{file}{Path to a file.}
}
\value{

}
\description{
Read a file into memory. We memory map the file for fast I/O.
The file is read in as a character vector of length = 1.
}
