% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastRead.R
\name{f_readlines}
\alias{f_readlines}
\title{Read lines from a file}
\usage{
f_readlines(file)
}
\arguments{
\item{file}{Path to a file.}
}
\value{
character vector
}
\description{
Read a file into memory. We memory map the file for fast I/O. The file is
read in as a character vector. This doesn't appear to scale. Large files do
not read into memory faster
}
