% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readMsp.R
\name{readMsp}
\alias{readMsp}
\title{Import a .msp file into R}
\usage{
readMsp(path)
}
\arguments{
\item{path}{system path to a .msp file containing mass spectral data}
}
\value{

}
\description{
Import a .msp file into R
}
\examples{

}
