% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readMsp.R
\name{.processHeader}
\alias{.processHeader}
\title{Process the header from an msp file}
\usage{
.processHeader(header)
}
\arguments{
\item{header}{character vector containing the header}
}
\value{
a named character vector
}
\description{
Process the header from an msp file
}
